# gh-125159: Speed up `posixpath.realpath()` in the common case