--- a/Misc/NEWS.d/next/Core_and_Builtins/2024-10-08-21-29-30.gh-issue-125159.jmi9FB.rst
+++ b/Misc/NEWS.d/next/Core_and_Builtins/2024-10-08-21-29-30.gh-issue-125159.jmi9FB.rst
@@ -1,0 +1,1 @@
+Speed up :func:`os.path.realpath` on Unix.