Originally proposed in #120146.

### Benchmark

<details><summary>Show script</summary>

```sh
# patch-1.sh
mkdir -p "tmp/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a"

echo 200 .
main/python.exe    -m timeit -s "import os; path = './' * 200"   "os.path.realpath(path)"
patch-1/python.exe -m timeit -s "import os; path = './' * 200"   "os.path.realpath(path)"

echo 200 ..
main/python.exe    -m timeit -s "import os; path = '../' * 200"   "os.path.realpath(path)"
patch-1/python.exe -m timeit -s "import os; path = '../' * 200"   "os.path.realpath(path)"

echo 200 dirs
main/python.exe    -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/a' * 200" "os.path.realpath(path)"
patch-1/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/a' * 200" "os.path.realpath(path)"
```

</details>

<details><summary>200 . - no difference</summary>

```none
10000 loops, best of 5: 26.9 usec per loop
10000 loops, best of 5: 26.7 usec per loop
```

</details>

<details><summary>200 .. - 1.02x faster </summary>

```none
5000 loops, best of 5: 50.3 usec per loop
5000 loops, best of 5: 49.2 usec per loop
```

</details>

<details><summary>200 dirs - no difference</summary>

```none
100 loops, best of 5: 2.19 msec per loop
100 loops, best of 5: 2.16 msec per loop
```

</details>

```shell
$ python -m timeit -s "name = '.'" "pass" "if name is None: pass"
50000000 loops, best of 5: 9.03 nsec per loop
$ python -m timeit -s "name = '.'" "pass"
50000000 loops, best of 5: 6.08 nsec per loop
```

<!-- gh-issue-number: gh-125159 -->
* Issue: gh-125159
<!-- /gh-issue-number -->