@nineteendo 2024-10-09
It's not a lot faster, 2.95 nsec/iteration.