@ericvsmith 2024-10-10
I think this change isn't worth the churn.