@erlend-aasland 2024-10-11
> Just as expected.

If I understand you correctly, you expected this PR to be rejected.

The next time you consider opening a PR that you "expect to be rejected", please consider again the following:

- is it worth the reviewer and CI churn to post a "ready-to-be-rejected" PR?
- is there anything you can improve _post_ proposing the PR that can increase the chances of it being accepted? (benchmarks, seeking consensus about the change, backwards compatibility, etc.) See also the devguide: [Making good PRs](https://devguide.python.org/getting-started/pull-request-lifecycle/#making-good-prs)