@barneygale 2024-10-11
@erlend-aasland it's my fault - see https://github.com/python/cpython/pull/120146#discussion_r1792511985

@nineteendo really sorry for giving you bad advice and wasting your time