@nineteendo 2024-10-11
Thanks for being honest, Barney, but it's also my fault. I knew it wasn't a lot faster (much less than the 5% Alex mentioned in the past) and therefore probably wasn't worth a PR to change just that. That doesn't mean that making the code 3 nanoseconds slower for speeding up a rare case is fine though.