@vstinner 2024-10-12
In general, an optimization is worth it if it's at least 10% faster on a microbenchmark. It doesn't seem to be the case here.